Tools and whatever

Tools and whatever

natural(vals, tolerance = .Machine$double.eps^0.5)

md5_checksum(df, ind)

user_widget()

version_info(newline = "<br>")

no_opt_out_ok()

insert_sample_data(include_data_table = TRUE)

delete_all_data(prompt = TRUE)

Arguments

vals

A numeric vector

tolerance

A (small) positive floating point number used to evaluate if a numeric can be regarded as a whole number. Default depends on the running environment and set to .Machine$double.eps^0.5

df

A data frame

ind

A data frame holding indicator data.

newline

String element defining line break for formatting. Default is <br>

include_data_table

Logical defining if the data table is to be populated by data too. By default TRUE

prompt

Logical to prompt for user input. Default is TRUE

Value

Invisible

Examples

version_info()
#> [1] "imongr v0.28.3<br>"
no_opt_out_ok()
#> [1] "Den er grei"